Replaced record card with record table
