+++ src/app/base/components/record-card/record-card.component.html
@@ -1,4 +1,4 @@
-<div class="px-4 py-5 sm:p-6">
+<!-- <div class="px-4 py-5 sm:p-6">
     <div *ngIf="recordCardOptions.recordComments[record.id]"
         class="tooltip tooltip-left cursor-pointer absolute top-6 right-5" data-tip="Record comments"
         (click)="recordCardOptions.dataBrowserModals.recordComments.open = true; recordCardOptions.dataBrowserModals.recordComments.commentsData = recordCardOptions.recordComments[record.id]">
@@ -92,4 +92,4 @@
             </span>
         </label>
     </div>
-</div>
+</div> -->
