+++ src/app/base/components/record-card/record-card.types.ts
@@ -1,38 +1,25 @@
-export type DataBrowserRecord = {
-    category: string;
-    created_at: string;
-    data: any;
-    db_order: number;
-    id: string;
-    project_id: string;
-    record_id: string;
-    rla_aggregation: any;
-    rla_data: any;
-    wsHint: string;
-};
+// export type DataBrowserRecord = {
+//     category: string;
+//     created_at: string;
+//     data: any;
+//     db_order: number;
+//     id: string;
+//     project_id: string;
+//     record_id: string;
+//     rla_aggregation: any;
+//     rla_data: any;
+//     wsHint: string;
+// };
 
-export type RecordDisplay = {
-    data: {
-        [k: string]: any;
-    };
-    [k: string]: any;
-};
 
-export type AttributeSort = {
-    key: string;
-    name: string;
-    order: number;
-    type: string;
-};
+// export type ActiveSearchParams = {
+//     id: number;
+//     searchText: string;
+//     searchTextReplaced: string;
+//     splittedText: string[];
+//     values: any;
+// };
 
-export type ActiveSearchParams = {
-    id: number;
-    searchText: string;
-    searchTextReplaced: string;
-    splittedText: string[];
-    values: any;
-};
-
-export type Attributes = {
-    [id: string]: any;
-}
+// export type Attributes = {
+//     [id: string]: any;
+// }
