+++ src/app/base/components/record-display/record-display.component.spec.ts
