+++ src/app/base/components/record-table/record-table.component.html
