+++ src/app/weak-supervision/components/weak-source-details/weak-source-details.component.ts
@@ -34,7 +34,7 @@ import { KnowledgeBasesApolloService } from 'src/app/base/services/knowledge-bas
 import { createDefaultHeuristicsDetailsModals, HeuristicsDetailsModals } from './weak-source-details-helper';
 import { AttributeVisibility } from 'src/app/projects/components/create-new-attribute/attributes-visibility-helper';
 import { BricksIntegratorComponent } from 'src/app/base/components/bricks-integrator/bricks-integrator.component';
-import { Attributes } from 'src/app/base/components/record-card/record-card.types';
+import { Attributes } from 'src/app/base/components/record-display/record-display.helper';
 
 @Component({
   selector: 'kern-weak-source-details',
